Moved all methods from checkstyle.api.utils to checkstyle.utils, issue #661
