+++ src/main/java/com/puppycrawl/tools/checkstyle/Checker.java
@@ -35,7 +35,6 @@ import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
 import com.puppycrawl.tools.checkstyle.api.MessageDispatcher;
 import com.puppycrawl.tools.checkstyle.api.SeverityLevel;
 import com.puppycrawl.tools.checkstyle.api.SeverityLevelCounter;
-import com.puppycrawl.tools.checkstyle.api.Utils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
