+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/AbstractFormatCheck.java
@@ -19,7 +19,7 @@
 package com.puppycrawl.tools.checkstyle.checks;
 
 import com.puppycrawl.tools.checkstyle.api.Check;
-import com.puppycrawl.tools.checkstyle.api.Utils;
+import com.puppycrawl.tools.checkstyle.Utils;
 
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
