+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/UncommentedMainCheck.java
@@ -22,7 +22,7 @@ import com.puppycrawl.tools.checkstyle.api.Check;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.FullIdent;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
-import com.puppycrawl.tools.checkstyle.api.Utils;
+import com.puppycrawl.tools.checkstyle.Utils;
 
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
