+++ src/main/java/com/puppycrawl/tools/checkstyle/filters/SuppressElement.java
@@ -20,7 +20,7 @@ package com.puppycrawl.tools.checkstyle.filters;
 
 import com.puppycrawl.tools.checkstyle.api.AuditEvent;
 import com.puppycrawl.tools.checkstyle.api.Filter;
-import com.puppycrawl.tools.checkstyle.api.Utils;
+import com.puppycrawl.tools.checkstyle.Utils;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
