+++ src/test/java/com/puppycrawl/tools/checkstyle/UtilsTest.java
@@ -22,8 +22,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import com.puppycrawl.tools.checkstyle.api.Utils;
-
 import java.io.File;
 import java.util.regex.Pattern;
